requirements.txt
pymongo~=4.8.0
Flask~=3.0.3
paho-mqtt~=2.1.0
pika~=1.3.2